+++ java/src/main/java/com/purescraps/purehtml/Validator.java
@@ -58,7 +58,7 @@ public class Validator {
                   },
                   "type": {
                     "type": "string",
-                    "enum": ["string", "object", "array", "union"]
+                    "enum": ["string", "number", "boolean", "object", "array", "union"]
                   },
                   "items": { "$ref": "config-schema.json" },
                   "properties": {